refactor(callers,workspace,deps): retarget worker imports to public module (#289)
Route every in-tree caller of the worker package through its new
public path at github.com/invakid404/baml-rest/worker, register the
worker module in the go workspace, and add the corresponding
require/replace pair to the parent module so the root build picks up
the local sources.

The CI workflow's subprocess test step and a handful of comments
that still pointed at the old internal/worker spelling are updated
in the same commit to keep checked-in source coherent.

The root module's embedded source manifest still references the old
paths; cmd/embed regen lands in the next commit.

Refs #289
