+++ .github/workflows/unit-tests.yml
@@ -58,7 +58,7 @@ jobs:
         # The default loop above runs everything under the no-tag build
         # (which is now the direct in-process worker path). The
         # `subprocess` tag flips a few key call sites and removes the
-        # recover wrappers in pool + internal/worker; this step
+        # recover wrappers in pool + worker; this step
         # exercises that the tag still compiles cleanly and the
         # subprocess-only tests pass. Kept narrow on purpose — running
         # -race -count=100 across every module under subprocess would
@@ -66,8 +66,8 @@ jobs:
         run: |
           go build ./...
           go build -tags=subprocess ./cmd/serve
-          go test -tags=subprocess -count=1 ./pool ./internal/worker ./cmd/serve
-          go test -race -tags=subprocess -count=1 ./pool ./internal/worker
+          go test -tags=subprocess -count=1 ./pool ./worker ./cmd/serve
+          go test -race -tags=subprocess -count=1 ./pool ./worker
 
       - name: Verify framework adapter codegen-emit
         # Two independent checks per adapter: deterministic emission
