+++ go.mod
@@ -14,6 +14,7 @@ require (
 	github.com/goccy/go-json v0.10.6
 	github.com/gofiber/contrib/v3/zerolog v1.0.6
 	github.com/gofiber/fiber/v3 v3.2.0
+	github.com/google/uuid v1.6.0
 	github.com/gregwebs/go-recovery v0.4.1
 	github.com/hashicorp/go-hclog v1.6.3
 	github.com/hashicorp/go-plugin v1.8.0
@@ -25,6 +26,7 @@ require (
 	github.com/invakid404/baml-rest/dynclient/baml-patched v0.0.0-00010101000000-000000000000
 	github.com/invakid404/baml-rest/introspected v0.0.0-00010101000000-000000000000
 	github.com/invakid404/baml-rest/pool v0.0.0-00010101000000-000000000000
+	github.com/invakid404/baml-rest/worker v0.0.0-00010101000000-000000000000
 	github.com/invakid404/baml-rest/workerplugin v0.0.0-00010101000000-000000000000
 	github.com/moby/buildkit v0.30.0
 	github.com/moby/moby/api v1.54.2
@@ -99,7 +101,6 @@ require (
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/protobuf v1.5.4 // indirect
 	github.com/google/shlex v0.0.0-20191202100458-e7afc7fbc510 // indirect
-	github.com/google/uuid v1.6.0 // indirect
 	github.com/gorilla/mux v1.8.0 // indirect
 	github.com/gregwebs/stackfmt v0.1.1 // indirect
 	github.com/grpc-ecosystem/grpc-gateway/v2 v2.28.0 // indirect
@@ -197,5 +198,6 @@ replace (
 	github.com/invakid404/baml-rest/dynclient/baml-patched => ./dynclient/baml-patched
 	github.com/invakid404/baml-rest/introspected => ./introspected
 	github.com/invakid404/baml-rest/pool => ./pool
+	github.com/invakid404/baml-rest/worker => ./worker
 	github.com/invakid404/baml-rest/workerplugin => ./workerplugin
 )
