+++ go.work
@@ -36,5 +36,6 @@ use (
 	./dynclient/baml-patched
 	./introspected
 	./pool
+	./worker
 	./workerplugin
 )
